+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## <> Changelog
+## Changelog
+
+### 1.2.0
+
+- changed "error without name" to "random 10 character name"
+
+### 1.1.0
+
+- switched to bash shell
 
 ### 1.0.0